+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - adding logs for shutdown handler of redis connections
+- `multiInstanceProcessing` allows to process the same event type/subtype on multiple application instances
 
 ### Fixed